requirements.txt
boto3~=1.21.42
botocore~=1.24.42
jmespath==0.10.0
pymongo~=4.1.1
PyMySQL~=1.0.2
python-dateutil==2.8.2
s3transfer==0.3.7
six==1.16.0
urllib3==1.26.9
PyYAML~=6.0